# specs/027-external-calendar-import/contracts/activity-draft.openapi.yaml
openapi: 3.1.0
info:
  title: Trip Planner — Activity Draft API
  version: 0.1.0
  description: |
    외부 캘린더에서 import된 draft 일정을 조회·승격·갱신·삭제한다.
    Spec: specs/027-external-calendar-import/spec.md
servers:
  - url: https://trip.idean.me/api
    description: production
  - url: https://dev.trip.idean.me/api
    description: dev
security:
  - sessionCookie: []
paths:
...
  /trips/{tripId}/drafts/{draftId}:
    delete:
      operationId: deleteDraft
      summary: draft 단건 삭제 (외부 캘린더에는 영향 없음)
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: string }
        - in: path
          name: draftId
          required: true
          schema: { type: string }
      responses:
        "204": { description: 삭제 완료 — 멱등 처리이므로 대상이 없어도 204 반환 }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
...